backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-allauth==0.57.0
psycopg2-binary==2.9.7
python-decouple==3.8
Pillow==10.0.1
celery==5.3.4
redis==5.0.1
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.14.1
django-extensions==3.2.3
django-filter==23.3
whitenoise==6.6.0
gunicorn==21.2.0